chore(generic-binding): move to generic_binding package
