+++ springwolf-add-ons/springwolf-generic-binding/src/main/java/io/github/stavshamir/springwolf/addons/generic_binding/annotation/configuration/SpringwolfGenericBindingConfiguration.java
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: Apache-2.0
-package io.github.stavshamir.springwolf.addons.genericbinding.annotation.configuration;
+package io.github.stavshamir.springwolf.addons.generic_binding.annotation.configuration;
 
-import io.github.stavshamir.springwolf.addons.genericbinding.annotation.processor.AsyncGenericOperationBindingProcessor;
+import io.github.stavshamir.springwolf.addons.generic_binding.annotation.processor.AsyncGenericOperationBindingProcessor;
 import io.github.stavshamir.springwolf.asyncapi.scanners.bindings.BindingProcessorPriority;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
