+++ springwolf-add-ons/springwolf-generic-binding/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,1 +1,1 @@
-io.github.stavshamir.springwolf.addons.genericbinding.annotation.processor.AsyncGenericOperationBindingProcessor
+io.github.stavshamir.springwolf.addons.generic_binding.annotation.processor.AsyncGenericOperationBindingProcessor
