+++ springwolf-add-ons/springwolf-generic-binding/src/test/java/io/github/stavshamir/springwolf/addons/generic_binding/annotation/processor/AsyncGenericOperationBindingProcessorTest.java
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: Apache-2.0
-package io.github.stavshamir.springwolf.addons.genericbinding.annotation.processor;
+package io.github.stavshamir.springwolf.addons.generic_binding.annotation.processor;
 
-import io.github.stavshamir.springwolf.addons.genericbinding.annotation.AsyncGenericOperationBinding;
+import io.github.stavshamir.springwolf.addons.generic_binding.annotation.AsyncGenericOperationBinding;
 import io.github.stavshamir.springwolf.asyncapi.scanners.bindings.ProcessedOperationBinding;
 import org.junit.jupiter.api.Test;
